Inheritance part 2

+++ section7/Inheritance/src/Animal.java
@@ -2,11 +2,11 @@ public class Animal {
 
     private String type;
     private String size;
-    private int weight;
+    private double weight;
 
     public Animal() {}
 
-    public Animal(String type, String size, int weight) {
+    public Animal(String type, String size, double weight) {
         this.type = type;
         this.size = size;
         this.weight = weight;
